chore: Move tests into right packages
